chore: cleanup project structure and fix clippy warnings
## Changes

### Code Quality (clippy fixes)
- hwp-core/export: Collapse nested if-else statement
- hwp-core/parser: Remove unnecessary type casts in tests
- hwp-mcp: Add #[allow(dead_code)] for future-use functions/fields

### Documentation
- CONTEXT.md: Update project structure, add hwp-wasm, update date
- CLAUDE.md: Update last modified date

### License
- Update all Cargo.toml files to use LicenseRef-Proprietary

### Configuration
- .gitignore: Add comprehensive patterns for Python, Node, IDE, logs

+++ crates/hwp-core/src/parser/docinfo/border_fill.rs
@@ -305,8 +305,8 @@ mod tests {
         data.extend_from_slice(&0u32.to_le_bytes()); // pattern_type
 
         // image info (5 bytes)
-        data.push(10i8 as u8); // brightness
-        data.push(0i8 as u8); // contrast
+        data.push(10u8); // brightness
+        data.push(0u8); // contrast
         data.push(0); // effect
         data.extend_from_slice(&5u16.to_le_bytes()); // bin_data_id
